+++ game_patch/object/event.cpp
@@ -256,8 +256,6 @@ ConsoleCommand2 debug_event_msg_cmd{
     "dbg_events",
     []() {
         event_debug_enabled = !event_debug_enabled;
-
-        //auto all_events = find_all_events_by_type(rf::EventType::Difficulty_Gate); // test
     }
 };
 
@@ -419,16 +417,93 @@ CodeInjection Event__turn_off_redirector_patch {
     }
 };
 
+void event_process_fixed_delay_patch(rf::GenericEvent* event) {
+
+    // fixed delay handling for events broken in stock game
+    if (af_rfl_version(rf::level.version) && event && event->delay_timestamp.elapsed()) {
+
+        if (event_debug_enabled) {
+            rf::console::print("Processing {} message in event {} ({}) (delayed)",
+                event->delayed_msg ? "ON" : "OFF", event->name, event->uid);
+        }
+
+        if (event->delayed_msg) {
+            switch (static_cast<rf::EventType>(event->event_type)) {
+                case rf::EventType::UnHide: {
+                    using EventTurnOnFn = void(__thiscall*)(rf::GenericEvent*);
+                    static EventTurnOnFn event_turn_on = reinterpret_cast<EventTurnOnFn>(0x004BCDD0);
+                    event_turn_on(event);
+                    break;
+                }
+                case rf::EventType::Alarm_Siren: {
+                    using EventTurnOnFn = void(__thiscall*)(rf::GenericEvent*);
+                    static EventTurnOnFn event_turn_on = reinterpret_cast<EventTurnOnFn>(0x004BA830);
+                    event_turn_on(event);
+                    break;
+                }
+                default: {
+                    xlog::warn("Something {} turned on", event->uid);
+                }
+            }
+        }
+        else {
+            switch (static_cast<rf::EventType>(event->event_type)) {
+                case rf::EventType::UnHide: {
+                    using EventTurnOnFn = void(__thiscall*)(rf::GenericEvent*);
+                    static EventTurnOnFn event_turn_off = reinterpret_cast<EventTurnOnFn>(0x004BCDE0);
+                    event_turn_off(event);
+                    break;
+                }
+                case rf::EventType::Alarm_Siren: {
+                    using EventTurnOnFn = void(__thiscall*)(rf::GenericEvent*);
+                    static EventTurnOnFn event_turn_off = reinterpret_cast<EventTurnOnFn>(0x004BA8C0);
+                    event_turn_off(event);
+                    break;
+                }
+                default: {
+                    xlog::warn("Something {} turned off", event->uid);
+                }
+            }
+        }
+
+        if (rf::event_type_forwards_messages(event->event_type)) {
+            using EventActivateLinksFn = void(__thiscall*)(rf::GenericEvent*, int, int, bool);
+            static EventActivateLinksFn event_activate_links = reinterpret_cast<EventActivateLinksFn>(0x004B8B00);
+            event_activate_links(event, event->trigger_handle, event->triggered_by_handle, event->delayed_msg);
+        }
+
+        event->delay_timestamp.invalidate();
+    }
+}
+
+CodeInjection EventUnhide__process_patch {
+    0x004BCDF0, [](auto& regs) {
+        rf::GenericEvent* event = regs.ecx;
+        event_process_fixed_delay_patch(event);
+    }
+};
+
+CodeInjection EventAlarmSiren__process_patch {
+    0x004BA880, [](auto& regs) {
+        rf::GenericEvent* event = regs.ecx;
+        event_process_fixed_delay_patch(event);
+    }
+};
+
 void apply_event_patches()
 {
-    // allow Holster_Player_Weapon and Holster_Weapon to be turned off (af levels)
+    // fix some events not working if delay value is specified (alpine levels only)
+    EventUnhide__process_patch.install();
+    EventAlarmSiren__process_patch.install();
+
+    // allow Holster_Player_Weapon and Holster_Weapon to be turned off (alpine levels only)
     Event__turn_off_redirector_patch.install();
 
-    // allow player flashlights via Headlamp_State (af levels)
+    // allow player flashlights via Headlamp_State (alpine levels only)
     event_headlamp_state_on_patch.install();
     event_headlamp_state_off_patch.install();
 
-    // allow event activation from triggers in multiplayer (af levels)
+    // allow event activation from triggers in multiplayer (alpine levels only)
     trigger_activate_linked_objects_patch.install();
 
     // Improve player teleport behaviour